refactor: retire the public backend seam and narrow crate surfaces
