+++ crates/loonfs-cli/Cargo.toml
@@ -27,6 +27,7 @@ loonfs-objectstore = { path = "../loonfs-objectstore" }
 serde.workspace = true
 serde_json.workspace = true
 tempfile.workspace = true
+thiserror.workspace = true
 tokio.workspace = true
 toml.workspace = true
 walkdir.workspace = true
